Remove extra triangles from the final triangulation
When we triangulate a set of points we compute a bounding
"supertriangle" that encloses all of them. This triangle isn't part of
the triangulation and shouldn't appear in the result.

For example, if we leave in the bounding triangle we use to compute the
triangulation of a square, we end up with 9 triangles in the result
(definitely wrong). Seven of these triangles share a vertex with the
bounding triangle and so can be removed, leaving the correct
triangulation behind.